+++ apps/antalmanac/src/components/RightPane/CoursePane/SearchForm/GeSelector.tsx
