+++ apps/antalmanac/src/components/RightPane/CoursePane/SearchForm/LegacySearch.tsx
@@ -3,7 +3,7 @@ import { Box, Button } from '@mui/material';
 import AdvancedSearch from '$components/RightPane/CoursePane/SearchForm/AdvancedSearch/AdvancedSearch';
 import CourseNumberSearchBar from '$components/RightPane/CoursePane/SearchForm/CourseNumberSearchBar';
 import { DepartmentSearchBar } from '$components/RightPane/CoursePane/SearchForm/DepartmentSearchBar/DepartmentSearchBar';
-import { GeSelector } from '$components/RightPane/CoursePane/SearchForm/GeSelectorFoo';
+import { GeSelector } from '$components/RightPane/CoursePane/SearchForm/GeSelector';
 import SectionCodeSearchBar from '$components/RightPane/CoursePane/SearchForm/SectionCodeSearchBar';
 
 interface LegacySearchProps {
